Convert JUnit 4 Parameterized test to @Nested tests in JUnit 5
